% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnacle.R
\name{implied_probs}
\alias{implied_probs}
\title{Calculate Implied Win Probabilities}
\usage{
implied_probs(df, id_col = "matchup_id", x = "venue", odds_col = "price")
}
\arguments{
\item{df}{A data frame.}

\item{id_col}{Column (quoted)  in \code{df} containing IDs.}

\item{x}{Column (quoted)  in \code{df} containing \code{id_col} groupings.}

\item{odds_col}{Column (quoted) in \code{df} containing odds (must be in decimal format).}
}
\description{
Calculate Implied Win Probabilities
}
