% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio.R
\name{odds_ratio}
\alias{odds_ratio}
\title{A simple function for calculating odds ratios from 2x2 tables}
\usage{
odds_ratio(a, b, c, d, ci = FALSE, alpha = 0.05, paired = FALSE)
}
\arguments{
\item{a}{The exposed + disease + count}

\item{b}{The exposed + disease - count}

\item{c}{The exposed - disease + count}

\item{d}{The exposed - disease - count}

\item{ci}{Return a confidence interval instead of point estimate}

\item{alpha}{The confidence level of the interval}

\item{paired}{Is the study design based on matched pairs?}
}
\description{
This function allows for odds ratio calculations.
}
\examples{
odds_ratio(30,40,10,50)
odds_ratio(30,40,10,50,ci = TRUE)
}
\keyword{odds-ratio}
