% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds_ratio_CMH.R
\name{odds_ratio_CMH}
\alias{odds_ratio_CMH}
\title{A function for calculating Cochran–Mantel–Haenszel odds ratios for across 2 strata}
\usage{
odds_ratio_CMH(a, b, c, d, ci = FALSE, alpha = 0.05, cor = FALSE)
}
\arguments{
\item{a}{A vector of the exposed + disease + count for each stratum}

\item{b}{A vector of the exposed + disease - count for each stratum}

\item{c}{A vector of the exposed - disease + count for each stratum}

\item{d}{A vector of the exposed - disease - count for each stratum}

\item{ci}{Return a confidence interval instead of point estimate}

\item{alpha}{The confidence level of the interval}

\item{cor}{Continuity correction}
}
\description{
This function allows for odds ratio calculations across two strata.
}
\examples{
odds_ratio(c(30,40),c(20,40),c(50,10),c(10,50))
}
\keyword{odds-ratio}
