% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnacle.R
\name{pinnacle_request}
\alias{pinnacle_request}
\title{Get Request to Pinnacle API}
\usage{
pinnacle_request(url, flatten = FALSE)
}
\arguments{
\item{url}{A character string.}

\item{flatten}{A logical passed to \code{httr2::resp_body_json()}.}
}
\value{
A data frame containing the extracted body from response
}
\description{
Get Request to Pinnacle API
}
\examples{
\dontrun{
# nhl endpoint
url <- https://guest.api.arcadia.pinnacle.com/0.1/leagues/1456

# make get request
pinnacle_request(url)
}

}
\keyword{internal}
